+++ src/__mocks__/transformers/babelTransformer.js
@@ -1,0 +1,16 @@
+const babelJest = require('babel-jest');
+
+module.exports = babelJest.createTransformer({
+  presets: [
+    ['@babel/preset-env', { targets: { node: 'current' } }],
+    '@babel/preset-typescript',
+    ['@babel/preset-react', { runtime: 'automatic' }]
+  ],
+  plugins: [
+    '@babel/plugin-transform-private-methods',
+    '@babel/plugin-transform-private-property-in-object',
+    '@babel/plugin-transform-class-properties'
+  ],
+  babelrc: false,
+  configFile: false,
+});